pyproject.toml
[project]
name = "cognition-mcp"
version = "0.1.0"
description = "MCP server for surfacing insights from your reading, viewing, and note-taking history"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.0.0",
    "httpx>=0.27.0",
    "feedparser>=6.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
cognition-mcp = "cognition_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cognition_mcp"]
